# Conflicts:
#	Modules/Voxel/voxel_generation.cpp

+++ Test/Test/test_cross_process_receive.cpp
@@ -35,6 +35,8 @@ namespace
     {
         char fallback_shared_memory_name[256];
 
+        if (::shm_unlink(shared_memory_name) == 0)
+            return (0);
         if (::unlink(shared_memory_name) == 0)
             return (0);
         if (errno == ENOENT || errno == EROFS)
